create_simple_mcp_servers.py
#!/usr/bin/env python3
"""
Create Simple MCP Model Servers
==============================

Creates simple model server files for all MCP models.
"""

import os
import logging

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

def create_simple_model_server(model_id, model_name, port):
    """Create a simple model server for a specific MCP model"""
    
    server_content = f'''"""
{model_name} MCP Model Server
============================

Simple model server for {model_name} MCP integration.
"""

import asyncio
import logging
import json
from datetime import datetime
from typing import Dict, Any, Optional
from fastapi import FastAPI, HTTPException
from pydantic import BaseModel
import uvicorn

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# FastAPI app
app = FastAPI(title="{model_name} MCP Server", version="1.0.0")

class MCPRequest(BaseModel):
    prompt: str
    context: Dict[str, Any]
    request_type: str
    priority: int = 1

class MCPResponse(BaseModel):
    response: str
    confidence: float
    tokens_used: int
    metadata: Dict[str, Any]

class MCPModel:
    """MCP model implementation for {model_name}"""
    
    def __init__(self):
        self.model_name = "{model_name}"
        self.usage_count = 0
        self.total_tokens = 0
        
    async def process_request(self, request: MCPRequest) -> MCPResponse:
        """Process an MCP request"""
        try:
            self.usage_count += 1
            
            # Generate response
            response_text = f"Response from {self.model_name}: Analysis of {{request.prompt[:100]}}..."
            confidence = 0.7 + (self.usage_count % 3) * 0.1
            tokens_used = len(request.prompt.split()) + 50
            self.total_tokens += tokens_used
            
            return MCPResponse(
                response=response_text,
                confidence=confidence,
                tokens_used=tokens_used,
                metadata={{
                    "model": self.model_name,
                    "timestamp": datetime.now().isoformat(),
                    "usage_count": self.usage_count,
                    "total_tokens": self.total_tokens
                }}
            )
            
        except Exception as e:
            logger.error(f"Error processing request: {{e}}")
            raise HTTPException(status_code=500, detail=str(e))

# Initialize model
model = MCPModel()

@app.get("/health")
async def health_check():
    """Health check endpoint"""
    return {{"status": "healthy", "model": "{model_name}", "timestamp": datetime.now().isoformat()}}

@app.post("/api/v1/chat", response_model=MCPResponse)
async def chat_endpoint(request: MCPRequest):
    """Chat endpoint"""
    return await model.process_request(request)

@app.post("/generate", response_model=MCPResponse)
async def generate_endpoint(request: MCPRequest):
    """Generate endpoint"""
    return await model.process_request(request)

@app.post("/completion", response_model=MCPResponse)
async def completion_endpoint(request: MCPRequest):
    """Completion endpoint"""
    return await model.process_request(request)

@app.post("/code", response_model=MCPResponse)
async def code_endpoint(request: MCPRequest):
    """Code endpoint"""
    return await model.process_request(request)

@app.post("/analyze", response_model=MCPResponse)
async def analyze_endpoint(request: MCPRequest):
    """Analyze endpoint"""
    return await model.process_request(request)

@app.post("/sentiment", response_model=MCPResponse)
async def sentiment_endpoint(request: MCPRequest):
    """Sentiment endpoint"""
    return await model.process_request(request)

@app.post("/risk", response_model=MCPResponse)
async def risk_endpoint(request: MCPRequest):
    """Risk endpoint"""
    return await model.process_request(request)

@app.get("/status")
async def get_status():
    """Get model status"""
    return {{
        "model": "{model_name}",
        "status": "active",
        "usage_count": model.usage_count,
        "total_tokens": model.total_tokens,
        "timestamp": datetime.now().isoformat()
    }}

if __name__ == "__main__":
    uvicorn.run(app, host="0.0.0.0", port={port})
'''
    
    with open(f"mcp_models/{model_id}/model_server.py", "w") as f:
        f.write(server_content)
    
    logger.info(f"  ✅ Created {model_name} server")

def main():
    """Create all MCP model servers"""
    logger.info("🖥️ Creating MCP model servers...")
    
    models = [
        ("gordon", "Gordon Assistant", 8001),
        ("llama2", "Llama2 7B", 8002),
        ("mistral", "Mistral 7B", 8003),
        ("codellama", "CodeLlama 7B", 8004),
        ("financial", "Financial LLM", 8005),
        ("sentiment", "Sentiment Analyzer", 8006),
        ("risk", "Risk Analyzer", 8007)
    ]
    
    for model_id, model_name, port in models:
        create_simple_model_server(model_id, model_name, port)
    
    logger.info("✅ All MCP model servers created!")

if __name__ == "__main__":
    main()
