requirements.txt
# Core AI and ML libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
torch==2.0.1
transformers==4.30.2
openai==0.27.8
scipy==1.11.1
# talib==0.4.25  # Commented out - requires special installation on Windows

# Trading and financial data
alpaca-trade-api==3.2.0
yfinance==0.2.18
alpha-vantage==2.3.1

# Web dashboard
flask==2.3.3
python-dotenv==1.0.0
ccxt==4.0.77

# Data processing and analysis
requests==2.31.0
beautifulsoup4==4.12.2
tweepy==4.14.0
textblob==0.17.1
vaderSentiment==3.3.2
huggingface_hub>=0.23.2
feedparser==6.0.10
pytz==2023.3

# Async and communication
asyncio-mqtt==0.13.0
redis==4.6.0
celery==5.3.1
fastapi==0.100.0
uvicorn==0.22.0

# Database and storage
sqlalchemy==2.0.19
psycopg2-binary==2.9.7

# Monitoring and logging
prometheus-client==0.17.1
structlog==23.1.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-mock==3.11.1

# Development tools
black==23.7.0
flake8==6.0.0
mypy==1.4.1

# Docker integration
docker==6.1.3
